Revert "Enable parallel SELECT for "INSERT INTO ... SELECT ..."."
To allow inserts in parallel-mode this feature has to ensure that all the
constraints, triggers, etc. are parallel-safe for the partition hierarchy
which is costly and we need to find a better way to do that. Additionally,
we could have used existing cached information in some cases like indexes,
domains, etc. to determine the parallel-safety.

List of commits reverted, in reverse chronological order:

2ed4061ddc Doc: Update description for parallel insert reloption.
5aefb7b163 Add a new GUC and a reloption to enable inserts in parallel-mode.
691a82c869 Improve FK trigger parallel-safety check added by 8ed57dfd65.
02a6ceb6a5 Fix use of relcache TriggerDesc field introduced by commit 8ed57dfd65.
988c351cc0 Fix valgrind issue in commit 8ed57dfd65.
8ed57dfd65 Enable parallel SELECT for "INSERT INTO ... SELECT ...".

Discussion: https://postgr.es/m/E1lMiB9-0001c3-SY@gemulon.postgresql.org